Exception handling

+++ src/main/java/com/sistempdv/pdv/service/ProductService.java
@@ -3,13 +3,16 @@ package com.sistempdv.pdv.service;
 import com.sistempdv.pdv.record.ProductRecord;
 import com.sistempdv.pdv.entity.Product;
 import com.sistempdv.pdv.repository.ProductRepository;
+import com.sistempdv.pdv.service.exception.DatabaseException;
+import com.sistempdv.pdv.service.exception.ResourceNotFoundException;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.List;
 
 @Service
 public class ProductService {
@@ -20,7 +23,7 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductRecord findById(Long id) {
-        Product product = productRepository.findById(id).get();
+        Product product = productRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Produto não encontrado"));
         return new ProductRecord(product);
     }
 
@@ -40,14 +43,27 @@ public class ProductService {
 
     @Transactional
     public ProductRecord update(Long id, ProductRecord productRecord){
-        Product entity = productRepository.getReferenceById(id);
-        copyRecordToEntity(productRecord, entity);
-        entity = productRepository.save(entity);
-        return new ProductRecord(entity);
+        try {
+            Product entity = productRepository.getReferenceById(id);
+            copyRecordToEntity(productRecord, entity);
+            entity = productRepository.save(entity);
+            return new ProductRecord(entity);
+        } catch (EntityNotFoundException e){
+            throw new ResourceNotFoundException("Produto não encontrado");
+        }
+
     }
 
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        productRepository.deleteById(id);
+        if (!productRepository.existsById(id)){
+            throw new ResourceNotFoundException("ID não encontrado!");
+        }
+        try {
+            productRepository.deleteById(id);
+        }catch (DataIntegrityViolationException e) {
+            throw new DatabaseException("Falha de integridade referencial!");
+        }
     }
